Move types to the types package
