+++ src/platform/packages/private/kbn-esql-editor/tsconfig.json
@@ -27,6 +27,7 @@
     "@kbn/code-editor",
     "@kbn/fields-metadata-plugin",
     "@kbn/esql-validation-autocomplete",
+    "@kbn/esql-variables-types",
     "@kbn/esql-utils",
     "@kbn/content-management-favorites-public",
     "@kbn/usage-collection-plugin",
