+++ src/platform/packages/shared/kbn-es-query/src/expressions/types.ts
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import type { ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import type { ESQLControlVariable } from '@kbn/esql-variables-types';
 import { Filter, Query, TimeRange } from '../filters';
 
 export interface ExecutionContextSearch {
