+++ src/platform/packages/shared/kbn-esql-utils/src/index.ts
@@ -22,9 +22,14 @@ export {
   retrieveMetadataColumns,
   getQueryColumnsFromESQLQuery,
   mapVariableToColumn,
+  getValuesFromQueryField,
 } from './utils/query_parsing_helpers';
 export { queryCannotBeSampled } from './utils/query_cannot_be_sampled';
-export { appendToESQLQuery, appendWhereClauseToESQLQuery } from './utils/append_to_query';
+export {
+  appendToESQLQuery,
+  appendWhereClauseToESQLQuery,
+  appendStatsByToQuery,
+} from './utils/append_to_query';
 export {
   getESQLQueryColumns,
   getESQLQueryColumnsRaw,
