+++ src/platform/packages/shared/kbn-esql-utils/src/utils/append_to_query.test.ts
@@ -7,7 +7,11 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { appendToESQLQuery, appendWhereClauseToESQLQuery } from './append_to_query';
+import {
+  appendToESQLQuery,
+  appendWhereClauseToESQLQuery,
+  appendStatsByToQuery,
+} from './append_to_query';
 
 describe('appendToQuery', () => {
   describe('appendToESQLQuery', () => {
@@ -173,6 +177,22 @@ and \`ip\`::string!="127.0.0.2/32"`
       expect(
         appendWhereClauseToESQLQuery('from logstash-*', 'dest', ['meow'], '+', 'string')
       ).toBeUndefined();
+    });
+  });
+
+  describe('appendStatsByToQuery', () => {
+    it('should append the stats by clause to the query', () => {
+      const queryString = 'FROM my_index';
+      const statsBy = 'my_field';
+      const updatedQueryString = appendStatsByToQuery(queryString, statsBy);
+      expect(updatedQueryString).toBe('FROM my_index\n| STATS BY my_field');
+    });
+
+    it('should append the stats by clause to the query with existing clauses', () => {
+      const queryString = 'FROM my_index | LIMIT 10 | STATS BY meow';
+      const statsBy = 'my_field';
+      const updatedQueryString = appendStatsByToQuery(queryString, statsBy);
+      expect(updatedQueryString).toBe('FROM my_index | LIMIT 10\n| STATS BY my_field');
     });
   });
 });
