+++ src/platform/packages/shared/kbn-esql-utils/src/utils/run_query.test.ts
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import { ESQLVariableType, type ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import { ESQLVariableType, type ESQLControlVariable } from '@kbn/esql-variables-types';
 import { getStartEndParams, getNamedParams } from './run_query';
 
 describe('run query helpers', () => {
