+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/index.ts
@@ -8,7 +8,6 @@
  */
 
 export type { SuggestionRawDefinition, ItemKind } from './src/autocomplete/types';
-export { ESQLVariableType, type ESQLControlVariable } from './src/shared/types';
 export { inKnownTimeInterval } from './src/shared/helpers';
 export type { CodeAction } from './src/code_actions/types';
 export type {
