+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/autocomplete/autocomplete.ts
@@ -8,6 +8,7 @@
  */
 
 import { uniq, uniqBy } from 'lodash';
+import type { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-variables-types';
 import type {
   AstProviderFn,
   ESQLAst,
@@ -89,12 +90,7 @@ import {
   getPolicyHelper,
   getSourcesHelper,
 } from '../shared/resources_helpers';
-import type {
-  ESQLCallbacks,
-  ESQLSourceResult,
-  ESQLControlVariable,
-  ESQLVariableType,
-} from '../shared/types';
+import type { ESQLCallbacks, ESQLSourceResult } from '../shared/types';
 import {
   getFunctionsToIgnoreForStats,
   getQueryForFields,
