+++ src/platform/packages/shared/kbn-esql-validation-autocomplete/src/shared/types.ts
@@ -6,6 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
+import type { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-variables-types';
 import type { ESQLRealField, JoinIndexAutocompleteItem } from '../validation/types';
 
 /** @internal **/
@@ -34,18 +35,6 @@ export interface ESQLSourceResult {
   title?: string;
   dataStreams?: Array<{ name: string; title?: string }>;
   type?: string;
-}
-
-export interface ESQLControlVariable {
-  key: string;
-  value: string | number;
-  type: ESQLVariableType;
-}
-
-export enum ESQLVariableType {
-  TIME_LITERAL = 'time_literal',
-  FIELDS = 'fields',
-  VALUES = 'values',
 }
 
 export interface ESQLCallbacks {
