+++ src/platform/packages/shared/kbn-esql-variables-types/kibana.jsonc
@@ -1,5 +1,5 @@
 {
-  "type": "shared-browser",
+  "type": "shared-common",
   "id": "@kbn/esql-variables-types",
   "owner": [
     "@elastic/kibana-esql",
