+++ src/platform/packages/shared/kbn-esql-variables-types/tsconfig.json
@@ -14,7 +14,6 @@
     "target/**/*",
   ],
   "kbn_references": [
-    "@kbn/esql-validation-autocomplete",
     "@kbn/presentation-publishing",
   ]
 }
