+++ src/platform/plugins/shared/controls/public/control_group/get_control_group_factory.tsx
@@ -9,7 +9,7 @@
 
 import { DataView } from '@kbn/data-views-plugin/common';
 import { ReactEmbeddableFactory } from '@kbn/embeddable-plugin/public';
-import { ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import { ESQLControlVariable } from '@kbn/esql-variables-types';
 import { PublishesESQLVariable, apiPublishesESQLVariable } from '@kbn/esql-variables-types';
 import { i18n } from '@kbn/i18n';
 import {
