+++ src/platform/plugins/shared/controls/public/controls/esql_control/esql_control_selections.ts
@@ -9,7 +9,7 @@
 import deepEqual from 'react-fast-compare';
 import { BehaviorSubject, combineLatest } from 'rxjs';
 import { PublishingSubject, StateComparators } from '@kbn/presentation-publishing';
-import { ESQLControlVariable, ESQLVariableType } from '@kbn/esql-validation-autocomplete';
+import { ESQLControlVariable, ESQLVariableType } from '@kbn/esql-variables-types';
 import type { ESQLControlState } from '@kbn/esql/public';
 
 export function initializeESQLControlSelections(initialState: ESQLControlState) {
