+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/control_flyout/field_control_form.test.tsx
@@ -11,7 +11,7 @@ import React from 'react';
 import { render, within, fireEvent } from '@testing-library/react';
 import { monaco } from '@kbn/monaco';
 import { dataPluginMock } from '@kbn/data-plugin/public/mocks';
-import { ESQLVariableType } from '@kbn/esql-validation-autocomplete';
+import { ESQLVariableType } from '@kbn/esql-variables-types';
 import { FieldControlForm } from './field_control_form';
 import { ESQLControlState, EsqlControlType } from '../types';
