+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/esql_control_action.ts
@@ -11,7 +11,7 @@ import { i18n } from '@kbn/i18n';
 import type { Action } from '@kbn/ui-actions-plugin/public';
 import type { CoreStart } from '@kbn/core/public';
 import type { ISearchGeneric } from '@kbn/search-types';
-import type { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import type { ESQLVariableType, ESQLControlVariable } from '@kbn/esql-variables-types';
 import { monaco } from '@kbn/monaco';
 import type { ESQLControlState } from './types';
