+++ src/platform/plugins/shared/esql/public/triggers/esql_controls/types.ts
@@ -6,7 +6,7 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import type { ESQLVariableType } from '@kbn/esql-validation-autocomplete';
+import type { ESQLVariableType } from '@kbn/esql-variables-types';
 
 export enum EsqlControlType {
   STATIC_VALUES = 'STATIC_VALUES',
