+++ src/platform/plugins/shared/esql/public/variables_service.ts
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import type { ESQLControlVariable } from '@kbn/esql-validation-autocomplete';
+import type { ESQLControlVariable } from '@kbn/esql-variables-types';
 
 export class EsqlVariablesService {
   esqlVariables: ESQLControlVariable[] = [];
